chore: move price in token to hook
